create put with id component to put in list item
